refactor(ingestion): move RenderModePayload to service module
Eliminates route-to-route import coupling by moving RenderModePayload
from datasets.py into render_mode.py, and adds a null-clear test for
the connections render-mode endpoint.
